Create new Game Component and restore Home for navigation
