+++ app/containers/GamePage.js
@@ -1,0 +1,10 @@
+import React, { Component } from 'react';
+import Game from '../components/Game';
+
+export default class GamePage extends Component {
+  render() {
+    return (
+      <Game />
+    );
+  }
+}
