+++ app/routes.js
@@ -2,12 +2,15 @@ import React from 'react';
 import { Route, IndexRoute } from 'react-router';
 import App from './containers/App';
 import HomePage from './containers/HomePage';
-import CounterPage from './containers/CounterPage';
+import GamePage from './containers/GamePage';
+import ParticleStormPage from './containers/ParticleStormPage';
 
 
 export default (
   <Route path="/" component={App}>
     <IndexRoute component={HomePage} />
-    <Route path="/counter" component={CounterPage} />
+    <Route path="/home" component={HomePage} />
+    <Route path="/game" component={GamePage} />
+    {/* <Route path="/particle-storm" component={ParticleStormPage} /> */}
   </Route>
 );
